requirements.txt
hydra-core==1.2.0
numpy==1.22.4
omegaconf==2.2.3
tqdm==4.63.0
torch==1.12.1